renames folder from token to tokenVoting
