+++ modules/client/src/tokenVoting/interfaces.ts
