+++ modules/client/src/tokenVoting/internal/graphql-queries/index.ts
