+++ modules/client/src/tokenVoting/internal/graphql-queries/proposal.ts
