+++ modules/client/src/tokenVoting/internal/graphql-queries/settings.ts
